# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OpenMuara
  description: Local-first billing and payment virtualization layer.
  version: 0.1.1
  contact:
    name: OpenMuara
servers:
  - url: http://127.0.0.1:9000
    description: Local development server
paths:
...
  /_admin/webhooks:
    get:
      summary: List recent webhook attempts
      operationId: listWebhooks
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, delivered, failed]
        - name: provider
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of webhook attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedWebhooks'
        '500':
          $ref: '#/components/responses/InternalError'
  /_admin/webhooks/replay-all:
    post:
      summary: Replay all webhook attempts matching optional filters
      operationId: replayAllWebhooks
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, delivered, failed]
        - name: provider
          in: query
          schema:
            type: string
      responses:
        '202':
          description: Number of queued replays
          content:
            application/json:
              schema:
                type: object
                properties:
                  replayed:
                    type: integer
  /_admin/webhooks/{ref}:
    get:
      summary: Inspect a webhook attempt
      operationId: inspectWebhook
      parameters:
        - name: ref
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Webhook attempt details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookAttempt'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      summary: Clear sensitive webhook payload data
      operationId: deleteWebhook
      parameters:
        - name: ref
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Data cleared
  /_admin/webhooks/{ref}/replay:
    post:
      summary: Replay a single webhook attempt
      operationId: replayWebhook
      parameters:
        - name: ref
          in: path
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Replay queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookAttempt'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    PaginatedWebhooks:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        results:
          type: array
          items:
            $ref: '#/components/schemas/WebhookAttempt'
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            trace_id:
              type: string
...
    WebhookAttempt:
      type: object
      properties:
        id:
          type: string
        ref:
          type: string
        provider_name:
          type: string
        url:
          type: string
        status:
          type: string
          enum: [pending, delivered, failed]
        attempts:
          type: integer
        last_error:
          type: string
        payload:
          type: string
          format: byte
        headers:
          type: object
          additionalProperties:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...